requirements.txt
Flask==2.3.2
Pillow==9.5.0
Pillow==10.0.0
SQLAlchemy==2.0.12
streamlit==1.22.0
